sqrt in sincos.ipp, and thomson function in test/mpi

+++ ReverseAD/include/reversead/reversead_base.hpp
@@ -2,16 +2,16 @@
 #define REVERSE_AD_BASE_H_
 
 #define NULL_LOC 0
-/*
+
 #define RANK_SHIFT 16
 #define RANK_BASE 0x000000000000fffful
 #define BASE_LOC 0x800000000000ul
-*/
 
+/*
 #define RANK_SHIFT 8
 #define RANK_BASE 0x000000ff
 #define BASE_LOC 0x800000
-
+*/
 #include "reversead/logger.hpp"
 
 namespace ReverseAD {
@@ -21,8 +21,8 @@ namespace ReverseAD {
   class TrivialTrace;
 
   typedef unsigned char opbyte;
-  typedef unsigned int locint;
-  //typedef unsigned long locint;
+  //typedef unsigned int locint;
+  typedef unsigned long locint;
 
   void logging_on();
   void trace_on();
